feat: Add confirmation prompts for pip installations.

+++ third-party.fish
@@ -1,11 +1,11 @@
 #!/usr/bin/fish
 source ./include/confirm.fish
 
-if type -q pip
+if type -q pip && confirm 'Do you want to install user-level software using pip?'
     pip install -r pkglist/requirements.txt
 end
 
-if sudo which pip >/dev/null
+if sudo which pip >/dev/null && confirm 'Do you want to install system-level software using pip?'
     sudo pip install -r pkglist/requirements-global.txt
 end
